requirements.txt
Django>=5.0,<6.0
django-cors-headers>=4.0,
django-filter>=24.3,<25.0,
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1